Enable `-Wconversion`
This commit replaces the existing `Wfloat-conversion` flag with the more
strict `-Wconversion` and fixes related errors. These were mostly
harmless integer conversions, but certainly also some nasty improper
floating point-to-integer conversions which may constitute bugs.

+++ device/common/include/traccc/device/impl/fill_prefix_sum.ipp
@@ -23,8 +23,10 @@ inline void fill_prefix_sum(
     }
 
     const prefix_sum_size_t previous =
-        (globalIndex == 0) ? 0 : sizes[globalIndex - 1];
-    const prefix_sum_size_t current = sizes[globalIndex];
+        (globalIndex == 0) ? 0
+                           : sizes[static_cast<unsigned int>(globalIndex - 1)];
+    const prefix_sum_size_t current =
+        sizes[static_cast<unsigned int>(globalIndex)];
     for (prefix_sum_size_t i = 0; i < current - previous; ++i) {
         result.at(previous + i) = {globalIndex, i};
     }
